# api/swagger.yaml
basePath: /
definitions:
  api.enqueueRequest:
    properties:
      id:
        type: integer
      type:
        type: string
    type: object
  model.Airing:
    properties:
      day:
        type: string
      end:
        $ref: '#/definitions/model.Date'
      start:
        $ref: '#/definitions/model.Date'
      time:
        type: string
    type: object
  model.AlternativeTitle:
    properties:
      english:
        type: string
      japanese:
        type: string
      synonym:
        type: string
    type: object
  model.Anime:
    properties:
      airing:
        $ref: '#/definitions/model.Airing'
      alternativeTitles:
        $ref: '#/definitions/model.AlternativeTitle'
      duration:
        type: string
      episode:
        type: integer
      favorite:
        type: integer
      genres:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      id:
        type: integer
      image:
        type: string
      licensors:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      member:
        type: integer
      popularity:
        type: integer
      premiered:
        type: string
      producers:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      rank:
        type: integer
      rating:
        type: integer
      related:
        $ref: '#/definitions/model.Related'
      score:
        type: number
      songs:
        $ref: '#/definitions/model.Song'
      source:
        type: integer
      status:
        type: integer
      studios:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      synopsis:
        type: string
      title:
        type: string
      type:
        type: integer
      video:
        type: string
      voter:
        type: integer
    type: object
  model.AnimeCharacter:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      role:
        type: string
      voiceActors:
        items:
          $ref: '#/definitions/model.Role'
        type: array
    type: object
  model.Character:
    properties:
      about:
        type: string
      favorite:
        type: integer
      id:
        type: integer
      image:
        type: string
      japaneseName:
        type: string
      name:
        type: string
      nicknames:
        items:
          type: string
        type: array
    type: object
  model.Date:
    properties:
      day:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
...
  model.EntryScore:
    properties:
      id:
        type: integer
      score:
        type: number
      title:
        type: string
      type:
        type: string
    type: object
  model.Item:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.Manga:
    properties:
      alternativeTitles:
        $ref: '#/definitions/model.AlternativeTitle'
      authors:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      chapter:
        type: integer
      favorite:
        type: integer
      genres:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      id:
        type: integer
      image:
        type: string
      member:
        type: integer
      popularity:
        type: integer
      publishing:
        $ref: '#/definitions/model.Publishing'
      rank:
        type: integer
      related:
        $ref: '#/definitions/model.Related'
      score:
        type: number
      serializations:
        items:
          $ref: '#/definitions/model.Item'
        type: array
      status:
        type: integer
      synopsis:
        type: string
      title:
        type: string
      type:
        type: integer
      volume:
        type: integer
      voter:
        type: integer
    type: object
...
  model.Publishing:
    properties:
      end:
        $ref: '#/definitions/model.Date'
      start:
        $ref: '#/definitions/model.Date'
    type: object
  model.Related:
    properties:
      adaptation:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      alternativeSetting:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      alternativeVersion:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      character:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      fullStory:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      other:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      parentStory:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      prequel:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      sequel:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      sideStory:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      spinOff:
        items:
          $ref: '#/definitions/model.Source'
        type: array
      summary:
        items:
          $ref: '#/definitions/model.Source'
        type: array
    type: object
  model.Role:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  model.Score:
    properties:
      "1":
        $ref: '#/definitions/model.ScoreDetail'
      "2":
        $ref: '#/definitions/model.ScoreDetail'
      "3":
        $ref: '#/definitions/model.ScoreDetail'
      "4":
        $ref: '#/definitions/model.ScoreDetail'
      "5":
        $ref: '#/definitions/model.ScoreDetail'
      "6":
        $ref: '#/definitions/model.ScoreDetail'
      "7":
        $ref: '#/definitions/model.ScoreDetail'
      "8":
        $ref: '#/definitions/model.ScoreDetail'
      "9":
        $ref: '#/definitions/model.ScoreDetail'
      "10":
        $ref: '#/definitions/model.ScoreDetail'
    type: object
  model.ScoreComparison:
    properties:
      anime:
        items:
          $ref: '#/definitions/model.EntryScore'
        type: array
      manga:
        items:
          $ref: '#/definitions/model.EntryScore'
        type: array
      novel:
        items:
          $ref: '#/definitions/model.EntryScore'
        type: array
    type: object
  model.ScoreDetail:
    properties:
      percent:
        type: number
      vote:
        type: integer
    type: object
  model.Song:
    properties:
      ending:
        items:
          type: string
        type: array
      opening:
        items:
          type: string
        type: array
    type: object
  model.Source:
    properties:
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  model.Stats:
    properties:
      score:
        $ref: '#/definitions/model.Score'
      summary:
        $ref: '#/definitions/model.Summary'
    type: object
  model.StatsHistory:
    properties:
      favorite:
        type: integer
      member:
        type: integer
      month:
        type: integer
      popularity:
        type: integer
      rank:
        type: integer
      score:
        type: number
      voter:
        type: integer
      year:
        type: integer
    type: object
  model.Summary:
    properties:
      completed:
        type: integer
      current:
        type: integer
      dropped:
        type: integer
      onHold:
        type: integer
      planned:
        type: integer
    type: object
...
  utils.Response:
    properties:
      data:
        type: object
      message:
        type: string
      meta:
        type: object
      status:
        type: integer
    type: object
info:
  contact:
    email: axel.rl.404@gmail.com
    name: Axel
    url: https://github.com/rl404
  description: MyAnimeList database dump and API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: MAL-DB API
paths:
  /anime/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Anime'
              type: object
      summary: Get anime details
      tags:
      - anime
  /anime/{id}/characters:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AnimeCharacter'
                  type: array
              type: object
      summary: Get anime character list
      tags:
      - anime
  /anime/{id}/staff:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get anime staff list
      tags:
      - anime
  /anime/{id}/stats:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Stats'
              type: object
      summary: Get anime stats
      tags:
      - anime
  /character/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Character'
              type: object
      summary: Get character details
      tags:
      - character
  /character/{id}/anime:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get character animeography list
      tags:
      - character
  /character/{id}/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get character mangaography list
      tags:
      - character
  /character/{id}/va:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get character voice actor list
      tags:
      - character
  /compare/score:
    get:
      consumes:
      - application/json
      parameters:
      - description: Manga title
        in: query
        name: title
        type: string
      - description: Order (negative means descending)
        enum:
        - title
        - -title
        - score
        - -score
        in: query
        name: order
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ScoreComparison'
                  type: array
              type: object
      summary: Entry stats history
      tags:
      - common
  /enqueue:
    post:
      consumes:
      - application/json
      parameters:
      - description: entry type and id
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.enqueueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Enqueue entry
      tags:
      - common
...
  /manga/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Manga'
              type: object
      summary: Get manga details
      tags:
      - manga
  /manga/{id}/characters:
    get:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Role'
                  type: array
              type: object
      summary: Get manga character list
      tags:
      - manga
  /manga/{id}/stats:
    get:
      consumes:
      - application/json
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Stats'
              type: object
      summary: Get manga stats
      tags:
      - manga
...
  /stats/history/{type}/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Entry type
        enum:
        - anime
        - manga
        - character
        - people
        in: path
        name: type
        required: true
        type: string
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.StatsHistory'
                  type: array
              type: object
      summary: Entry stats history
      tags:
      - common
...
schemes:
- http
- https
swagger: "2.0"